feat: Scheduling Pagfe | logic

+++ FrontEnd/barberschedule_client/lib/features/home/domain/models/order_model.dart
@@ -8,7 +8,7 @@ class OrderModel {
   final String barberShopId;
   final String dateTime;
   final String paymentMethodId;
-  final int price;
+  final double price;
   final String status;
   final BarberShopModel? barberShop;
   OrderModel({
@@ -41,7 +41,7 @@ class OrderModel {
       barberShopId: map['barberShopId'] as String,
       dateTime: map['dateTime'] as String,
       paymentMethodId: map['paymentMethodId'] as String,
-      price: map['price'] as int,
+      price: map['price'] as double,
       status: map['status'] as String,
       barberShop: map['barberShop'] == null
           ? null
